backend/requirements.txt
alembic==1.11.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
PyYAML==6.0.3
rich==13.9.4
ruff==0.14.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
pwdlib[argon2]==0.3.0
argon2-cffi
python-multipart
